Move bidictionary to common location
