+++ app/MonsterConverterInterface/BiDictionary.cs
@@ -7,7 +7,7 @@ using System.Runtime.Serialization;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MonsterConverterTfsXml
+namespace MonsterConverterInterface
 {
     [Serializable]
     [DebuggerDisplay("Count = {Count}"), DebuggerTypeProxy(typeof(DictionaryDebugView<,>))]
